Huy4

+++ server/config/services.php
@@ -38,7 +38,7 @@ return [
         'api_key' => env('BREVO_API_KEY'),
         'sender_name' => env('BREVO_SENDER_NAME'),
         'sender_email' => env('BREVO_SENDER_EMAIL'),
-        'otp_valid_minutes' => env('OTP_VALID_MINUTES')
+        'otp_valid_minutes' => (int) env('OTP_VALID_MINUTES')
     ],
 
     'ghn' => [
